+++ .kitchen.yml
@@ -4,7 +4,6 @@ driver:
 provisioner:
   name: chef_zero
   require_chef_omnibus: 12.5.1
-  data_path: test/shared
 
 platforms:
   - name: ubuntu-12.04
