+++ test/integration/languages_erlang/serverspec/languages_erlang_spec.rb
@@ -1,8 +1,4 @@
-require 'serverspec'
-require 'pathname'
-require 'tmpdir'
-
-require_relative 'spec_helper'
+require 'spec_helper'
 
 describe file '/opt/languages/erlang/18.1/releases/18/OTP_VERSION' do
   it { should exist }
